requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
python-docx
PyMuPDF
